+++ src/index.js
@@ -1,5 +1,4 @@
-//const { semanticReleaseConfig } = require('./config');
-const { merge } = require('./utils');
+const { merge, openJsonFile } = require('./utils');
 const { configureBranches } = require('./branches');
 const { configureTagFormat } = require('./tagFormat');
 const { configurePlugins } = require('./plugins');
@@ -29,6 +28,31 @@ const semanticReleaseConfigDocker = (options = {}) => {
   )
 }
 
+const semanticReleaseConfigDockerMulti = (options = {}) => {
+  if (!options.projectPath) {
+    throw 'Invalid Docker Multi config. Must set projectPath param in any docker project\'s release config.\nSee README."'
+  }
+  const packageJson = openJsonFile(`${options.projectPath}/package.json`)
+  const overrides = {
+    docker: true,
+    dockerMulti: true,
+    plugins: {
+      git: {
+        skipCi: false
+      },
+      docker: {
+        dockerTags: [
+          `{{#if prerelease.[0]}}${packageJson.docker_version}-{{prerelease.[0]}}{{else}}${packageJson.docker_version}-latest{{/if}}`,
+          `${packageJson.docker_version}-{{version}}`
+        ]
+      }
+    }
+  }
+  return semanticReleaseConfigDefault(
+    merge(merge(defaultOptions, overrides), options)
+  )
+}
+
 const semanticReleaseConfigTerraform = (options = {}) => {
   const overrides = {
     terraform: true,
@@ -46,5 +70,6 @@ const semanticReleaseConfigTerraform = (options = {}) => {
 module.exports = {
   semanticReleaseConfigDefault,
   semanticReleaseConfigDocker,
-  semanticReleaseConfigTerraform
+  semanticReleaseConfigTerraform,
+  semanticReleaseConfigDockerMulti
 }