+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [2.0.3-issue-docker-multi.3](https://gitlab.com/beepbeepgo/public/libraries/npm/nodejs-semantic-release-config/compare/2.0.3-issue-docker-multi.2...2.0.3-issue-docker-multi.3) (2022-12-16)
+
+
+### Bug Fixes
+
+* parse package json properly ([b75e00d](https://gitlab.com/beepbeepgo/public/libraries/npm/nodejs-semantic-release-config/commit/b75e00d355b456ba41d6df94a9da522762c88059))
+
+## [2.0.3-issue-docker-multi.2](https://gitlab.com/beepbeepgo/public/libraries/npm/nodejs-semantic-release-config/compare/2.0.3-issue-docker-multi.1...2.0.3-issue-docker-multi.2) (2022-12-16)
+
+
+### Bug Fixes
+
+* set packageJson to right value ([f6b9a29](https://gitlab.com/beepbeepgo/public/libraries/npm/nodejs-semantic-release-config/commit/f6b9a29c51606e03671a6a19367ac0b4ff8f0f23))
+
+## [2.0.3-issue-docker-multi.1](https://gitlab.com/beepbeepgo/public/libraries/npm/nodejs-semantic-release-config/compare/2.0.2...2.0.3-issue-docker-multi.1) (2022-12-16)
+
+
+### Bug Fixes
+
+* add docker multi support ([a3a4529](https://gitlab.com/beepbeepgo/public/libraries/npm/nodejs-semantic-release-config/commit/a3a4529f6d16d343e1d6fb7fe5c3bd1ba0d3c714))
+
 ## [2.0.2](https://gitlab.com/beepbeepgo/public/libraries/npm/nodejs-semantic-release-config/compare/2.0.1...2.0.2) (2022-12-16)